flashlight: Add support for simpler sysfs toggles (#34)
Unlike the overcomplicated Qualcomm mess, there are also simpler sysfs
toggles which you can just write 1/0 to when enabling/disabling the
flash.

This is used by the MediaTek flashlight_core driver on the Volla Phone
(Helio P23 / MT6763) for example, which this also adds support for. Pinephone
also uses a simple sysfs toggle for flashlight.